chore(vite-dev-server): rename waitForSupportFile.ts to warmupTargets.ts
The file no longer contains the waitUntilUrlReady polling helper that
gave it its original name; what's left is two URL/path builders
(getSupportFileRelativePath and getSpecRelativeUrl) used by devServer.ts
to construct the targets passed to server.warmupRequest. Rename to
match what's actually inside.

No behavior change.
